Fix warnings found when building with MSVC
This fixes all the warnings found in the library when building with MSVC
2017 for x86.

+++ server/MemMap.cpp
@@ -58,7 +58,7 @@ MemMap::Type MemMap::findRegion(const uint64_t addr,
 
   if (size > 0)
     for (auto it = mMemMap.begin(); it != mMemMap.end(); it++) {
-      const uint32_t locAddr = (addr + size - 1);
+      const uint64_t locAddr = (addr + size - 1);
 
       if ((it->start <= locAddr) && (locAddr <= it->end)) {
         endType = it->type;
